Add support external command files
If a commands string starts with File: the commands/cheats will be
read from the file path that is specified.
The external command file uses a format where each command/cheat is
placed on its own line.

This is intended to facilitate automation scenarios, e.g. automatically
placing networks and zones when a city is established.