Directly call cISC4ZoneManager::PlaceZone
This allows for better error reporting and reduces the load on the
game's messaging system when large numbers of zones are being placed.

Implemented the SC4CellRegion and cRZCellMap classes.
The cRZCellMap class is a template in SC4, but our version doesn't do
that because the only template parameter the game's code uses is bool.