Add a ZoneBitmap auto-run command
This command loads a bitmap containing a zone layout, and applies it
when loading a city.

The zone bitmap is a RGB image with different colored pixels for the
various zone types.
The bitmap has one more row than that city size because the bottom row
is used to allow the user to specify which pixel colors map to each
zone and control various options like placing streets.
By default streets are not placed because the main use case of this
feature is likely to be users laying down dummy zones as markers for
the city transportation networks.

The bitmap is loaded using SC4's built-in Gimex library, this allows
both BMP and PNG images to be used.

The DLL will write a text file with the zone colors next to the image
file, this simplifies the process of adding an override to make the
game's zone colors match the image.