Fix the LoadZoneBitmap cheat for paths with spaces
Using StringViewUtil::Split would cause the file to not be found if the
path contained spaces without using quotes.

As the only LoadZoneBitmap cheat parameter is the file name, it is more
user friendly to strip the cheat name from the string than to require
quotes for paths with spaces.
If the user includes quotes in the path they will be removed because
the OS doesn't support processing a quoted path.